+++ servers/Azure.Mcp.Server/CHANGELOG.md
@@ -7,6 +7,9 @@ The Azure MCP Server updates automatically by default whenever a new release com
 ### Features Added
 
 - Added support for `azmcp sql server list` command to list SQL servers in a subscription and resource group. [[#503](https://github.com/microsoft/mcp/issues/503)]
+- Added the following Azure Managed Lustre commands: 
+  - `azmcp_azuremanagedlustre_filesystem_get_sku_info`: Get information about Azure Managed Lustre SKU [[#100](https://github.com/microsoft/mcp/issues/100)].
+  - `azmcp_azuremanagedlustre_filesystem_subnetsize_validate`: Check if the subnet can host the target Azure Managed Lustre SKU and size [[#110](https://github.com/microsoft/mcp/issues/110)].
 
 ### Breaking Changes
 
@@ -21,6 +24,8 @@ The Azure MCP Server updates automatically by default whenever a new release com
 ### Bugs Fixed
 
 ### Other Changes
+- Changed the following Azure Managed Lustre commands:
+  - `azmcp_azuremanagedlustre_filesystem_required-subnet-size` becomes `azmcp_azuremanagedlustre_filesystem_subnetsize_ask`
 
 ## 0.8.0 (2025-09-18)